refactor: data folder organization
