+++ data/catalog/_data_main/__init__.py
@@ -1,4 +1,7 @@
 import os
 import sys
 
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+current_dir = os.path.dirname(os.path.realpath(__file__))
+sys.path.append(current_dir)
+sys.path.append(os.path.join(current_dir, "raw_data"))
+sys.path.append(os.path.join(current_dir, "process_data"))
