--- data/catalog/_data_main/process_boston_housing_data.py
@@ -1,8 +1,0 @@
-import pandas as pd
-
-
-def load_boston_housing_data():
-    file_path = "boston_housing.csv"
-    data_frame = pd.read_csv(file_path)
-
-    return data_frame
