+++ data/catalog/_data_main/process_data/process_compas_data.py
@@ -5,9 +5,9 @@ import urllib.request
 from collections import defaultdict
 from random import seed, shuffle
 
-import fair_utils_data as ut
 import numpy as np
 import pandas as pd
+import process_data.process_utils_data as ut
 from sklearn import preprocessing
 
 SEED = 1234
